Async rework 2.0 (#50)
* make apsManager and openAPS async....highly wip

* cleanup

* clean up

* test

* cleanup

* small fixes for LA, refactoring

* refactoring of async functions

* cleanup

* async save function for Determination

* address PR feedback (better error handling)

---------

+++ FreeAPS/Sources/Models/Determination.swift
@@ -7,7 +7,7 @@ struct Determination: JSON, Equatable {
     let eventualBG: Int?
     let sensitivityRatio: Decimal?
     let rate: Decimal?
-    let duration: Int?
+    let duration: Decimal?
     let iob: Decimal?
     let cob: Decimal?
     var predictions: Predictions?
